--- a/tests/src/bench.rs
+++ b/tests/src/bench.rs
@@ -1,8 +1,8 @@
-extern crate test;
 extern crate tempdir;
+extern crate test;
 
 use tempdir::TempDir;
-use harness::{Harness, Autojumper, Shell, Pazi, Fasd, NoJumper};
+use harness::{Autojumper, Fasd, Harness, NoJumper, Pazi, Shell};
 use self::test::Bencher;
 
 fn cd_bench(b: &mut Bencher, jumper: &Autojumper, shell: &Shell) {
@@ -22,11 +22,7 @@ fn cd_bench(b: &mut Bencher, jumper: &Autojumper, shell: &Shell) {
     let mut iter = 0;
 
     b.iter(move || {
-        let dir = if iter % 2 == 0 {
-            &dir1
-        } else {
-            &dir2
-        };
+        let dir = if iter % 2 == 0 { &dir1 } else { &dir2 };
         iter += 1;
         h.visit_dir(dir)
     });
@@ -69,25 +65,23 @@ fn jump_large_db_bench(b: &mut Bencher, jumper: &Autojumper, shell: &Shell) {
     });
 }
 
-
 fn cd_50_bench(b: &mut Bencher, jumper: &Autojumper, shell: &Shell) {
     let tmpdir = TempDir::new("pazi_bench").unwrap();
     let root = tmpdir.path();
     let mut h = Harness::new(&root, jumper, shell);
-    let dirs = (0..50).map(|num| {
-        format!("{}/dir{}", root.to_str().unwrap(), num)
-    }).collect::<Vec<_>>();
+    let dirs = (0..50)
+        .map(|num| format!("{}/dir{}", root.to_str().unwrap(), num))
+        .collect::<Vec<_>>();
     for dir in &dirs {
         h.create_dir(&dir);
     }
 
-    let cmd = dirs.iter().map(|el| {
-        format!("cd '{}'", el)
-    }).collect::<Vec<_>>().join(" && ");
+    let cmd = dirs.iter()
+        .map(|el| format!("cd '{}'", el))
+        .collect::<Vec<_>>()
+        .join(" && ");
 
-    b.iter(move || {
-        h.run_cmd(&cmd)
-    });
+    b.iter(move || h.run_cmd(&cmd));
 }
 
 // This file is generated with 'build.rs' based on the contents of 'src/benches.csv'; to change the