--- a/tests/src/harness/mod.rs
+++ b/tests/src/harness/mod.rs
@@ -41,8 +41,7 @@ impl Harness {
     }
 
     pub fn jump(&mut self, search: &str) -> String {
-        self.testshell
-            .run(&format!("z '{}' && pwd", search))
+        self.testshell.run(&format!("z '{}' && pwd", search))
     }
 
     pub fn run_cmd(&mut self, cmd: &str) -> String {