--- a/tests/src/harness/shells.rs
+++ b/tests/src/harness/shells.rs
@@ -7,8 +7,7 @@ use std::process::Command;
 pub enum Shell {
     Bash,
     Zsh,
-    #[allow(dead_code)]
-    Conch,
+    #[allow(dead_code)] Conch,
 }
 
 impl Shell {