🔥 🚗 ⏩ ⏩ 💨 

Before:
```
$ cargo bench --features=nightly cd_bench_nojumper
running 2 tests
test bench::cd_bench_nojumper_bash        ... bench:   5,886,133 ns/iter (+/- 3,493,799)
test bench::cd_bench_nojumper_zsh         ... bench:  14,084,141 ns/iter (+/- 9,686,470)
```

After:
```
$ cargo bench --features=nightly cd_bench_nojumper
running 2 tests
test bench::cd_bench_nojumper_bash        ... bench:      86,638 ns/iter (+/- 4,373)
test bench::cd_bench_nojumper_zsh         ... bench:      32,791 ns/iter (+/- 1,888)
```

Sorry about throwing in the noise of `cargo fmt` too, just ignore that commit for reviewing. The first two are the important ones that make real changes.