@euank 2017-12-20
Failures are because `rustup` install is 404-ing